# backend/public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Ficha de Treino API",
    "version": "0.1.0",
    "description": "API REST do sistema Ficha de Treino. Endpoints de autenticação, treinos, exercícios e gerenciamento de conta."
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Servidor da API"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/workouts": {
      "get": {
        "tags": [
          "Workouts"
        ],
        "summary": "Lista os 7 treinos semanais do usuário autenticado",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Treinos do usuário autenticado com contagem e nomes de exercícios",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "workouts": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "weekDay": {
                            "type": "string",
                            "enum": [
                              "DOMINGO",
                              "SEGUNDA",
                              "TERCA",
                              "QUARTA",
                              "QUINTA",
                              "SEXTA",
                              "SABADO"
                            ]
                          },
                          "exerciseCount": {
                            "type": "integer"
                          },
                          "exercises": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "required": [
                                "name",
                                "done"
                              ],
                              "properties": {
                                "name": {
                                  "type": "string"
                                },
                                "done": {
                                  "type": "boolean"
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token JWT ausente, inválido ou expirado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Usuário autenticado não encontrado no banco",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/workouts/{weekDay}": {
      "get": {
        "tags": [
          "Workouts"
        ],
        "summary": "Lista os exercícios de um treino diário do usuário autenticado",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "weekDay",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "DOMINGO",
                "SEGUNDA",
                "TERCA",
                "QUARTA",
                "QUINTA",
                "SEXTA",
                "SABADO"
              ]
            },
            "description": "Dia da semana oficial do treino"
          }
        ],
        "responses": {
          "200": {
            "description": "Treino diário com exercícios completos ordenados por nome",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "workout": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "weekDay": {
                          "type": "string",
                          "enum": [
                            "DOMINGO",
                            "SEGUNDA",
                            "TERCA",
                            "QUARTA",
                            "QUINTA",
                            "SEXTA",
                            "SABADO"
                          ]
                        },
                        "exercises": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "integer"
                              },
                              "done": {
                                "type": "boolean"
                              },
                              "exercise": {
                                "type": "object",
                                "properties": {
                                  "id": {
                                    "type": "string"
                                  },
                                  "isCustom": {
                                    "type": "boolean"
                                  },
                                  "name": {
                                    "type": "string"
                                  },
                                  "force": {
                                    "type": "string",
                                    "nullable": true
                                  },
                                  "level": {
                                    "type": "string"
                                  },
                                  "mechanic": {
                                    "type": "string",
                                    "nullable": true
                                  },
                                  "equipment": {
                                    "type": "string",
                                    "nullable": true
                                  },
                                  "primaryMuscles": {
                                    "type": "array",
                                    "items": {
                                      "type": "string"
                                    }
                                  },
                                  "secondaryMuscles": {
                                    "type": "array",
                                    "items": {
                                      "type": "string"
                                    }
                                  },
                                  "instructions": {
                                    "type": "array",
                                    "items": {
                                      "type": "string"
                                    }
                                  },
                                  "category": {
                                    "type": "string"
                                  },
                                  "images": {
                                    "type": "array",
                                    "items": {
                                      "type": "string"
                                    }
                                  }
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token JWT ausente, inválido ou expirado"
          },
          "404": {
            "description": "Dia inválido, treino inexistente ou treino de outro usuário"
          }
        }
      }
    },
    "/api/workouts/{weekDay}/exercises": {
      "post": {
        "tags": [
          "Workouts"
        ],
        "summary": "Adiciona um exercício ao treino diário do usuário autenticado",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "weekDay",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "DOMINGO",
                "SEGUNDA",
                "TERCA",
                "QUARTA",
                "QUINTA",
                "SEXTA",
                "SABADO"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "exerciseId"
                ],
                "properties": {
                  "exerciseId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Associação criada com done inicialmente falso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "exerciseId",
                    "done"
                  ],
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "exerciseId": {
                      "type": "string"
                    },
                    "done": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Corpo da requisição inválido"
          },
          "401": {
            "description": "Token JWT ausente, inválido ou expirado"
          },
          "404": {
            "description": "Dia, treino ou exercício não encontrado"
          },
          "409": {
            "description": "Exercício já existe no treino"
          }
        }
      }
    },
    "/api/workouts/{weekDay}/exercises/{exerciseId}": {
      "delete": {
        "tags": [
          "Workouts"
        ],
        "summary": "Remove um exercício do treino diário do usuário autenticado",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "weekDay",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "DOMINGO",
                "SEGUNDA",
                "TERCA",
                "QUARTA",
                "QUINTA",
                "SEXTA",
                "SABADO"
              ]
            }
          },
          {
            "in": "path",
            "name": "exerciseId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Identificador do exercício no catálogo"
          }
        ],
        "responses": {
          "200": {
            "description": "Exercício removido",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "deleted"
                  ],
                  "properties": {
                    "deleted": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token JWT ausente, inválido ou expirado"
          },
          "404": {
            "description": "Dia inválido, associação inexistente ou pertencente a outro usuário"
          }
        }
      }
    },
    "/api/workouts/{weekDay}/clear": {
      "post": {
        "tags": [
          "Workouts"
        ],
        "summary": "Desmarca todos os exercícios de um treino diário",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "weekDay",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "DOMINGO",
                "SEGUNDA",
                "TERCA",
                "QUARTA",
                "QUINTA",
                "SEXTA",
                "SABADO"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Quantidade de associações desmarcadas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "cleared"
                  ],
                  "properties": {
                    "cleared": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token JWT ausente, inválido ou expirado"
          },
          "404": {
            "description": "Dia inválido ou treino inexistente"
          }
        }
      }
    },
...
  },
...
}